Move tests to correct place
